Make it safe for `AdminRealmListener::start` to be called multiple times
`AdminRealmListener::start` is called each time `addListener` is called
from JavaScript, so it needs to correctly handle this situation. It was
very close to handling it correctly, with the exception of moving
`m_config` when opening the Realm which resulted in later calls dealing
with a `Realm::Config` instance in an indeterminate state, leading to
surprising exceptions and/or crashes.