Report sync errors which occur within the global notifier and data adator (#395)
* Report sync errors which occur within the global notifier and data adaptor

Any errors which occurred when trying to sync the admin Realm were ignored,
which made attempting to add a listener with an invalid admin user silently do
nothing.

* Report sync errors which occur within the global notifier and data adaptor

Any errors which occurred when trying to sync the admin Realm were ignored,
which made attempting to add a listener with an invalid admin user silently do
nothing.

* Update to include realm-js #1616